Add Maven repository and update writeNBT method
Updated the build.gradle file to include a specific Maven repository for publishing. This is to ensure that the correct repo is used during the publishing process.

Also changed the writeNBT method in ScaledSectionStorage.kt to take an additional parameter of type LevelChunkSection. This change will allow for more persistent storage, improving the overall durability and reliability of our software.